✨ (nds-casting) Cast NDS to uint256 to allow more sorting and reduce gas

+++ contracts/compound/positions-manager-parts/PositionsManagerGovernance.sol
@@ -31,7 +31,7 @@ abstract contract PositionsManagerGovernance is PositionsManagerEventsErrors {
 
     /// @notice Sets `NDS`.
     /// @param _newNDS The new `NDS` value.
-    function setNDS(uint8 _newNDS) external onlyOwner {
+    function setNDS(uint256 _newNDS) external onlyOwner {
         NDS = _newNDS;
         emit NDSSet(_newNDS);
     }
